feat: 공유버튼 모바일과 데탑 분리
