+++ client/src/components/ShareEventButton/index.ts
@@ -1,1 +1,2 @@
-export {default as ShareEventButton} from './ShareEventButton';
+export {default as DesktopShareEventButton} from './DesktopShareEventButton';
+export {default as MobileShareEventButton} from './MobileShareEventButton';
